Forbid SCONE packet insertion
This is one potential approach to #21.  However, it does constrain things more
than I think is necessary.  I tend to think that staying quiet is best,
because it gives network elements the option and endpoints the choice of
ignoring those inserted SCONE packets.  I wouldn't want to *encourage*
that outcome though.

Closes #21.