Add an indication
This adds a small indicator to the end of packets, so that network
elements can maybe flip their posture for those flows.
That requires a bit of a gamble, because the indication doesn't mean
that SCONE is supported or that throughput advice will be followed, but
it seems like we have some belief that this will change how the protocol
might be deployed.

I didn't strongly mandate the new flow indication piece.  That's
something we could do, but it turned out to be incredibly hard to
specify something that works sensibly.  Tying this to
PATH_CHALLENGE/PATH_RESPONSE does work, but those have more uses than
initializing a flow on a new path, so the set of conditions you need to
set out for when the indication is attached to a datagram end up being
very complicated.

Given that we already have good reason to send SCONE packets on new
paths, I don't think we need to overthink that part.

Closes #5.
Closes #23.