Update the UDP checksum
This might seem obvious to some of us, but it's an easy thing to
overlook when implementing.

I don't want to mention the all-zero UDP checksum thing.  I'm sure that
anyone looking into this will discover that and be able to reach their
own conclusions about how feasible that is.

Closes #42.