Monitoring and time changes (#61)
* Add a time component of 60 seconds

Closes #20.

* Add excerpt to gitignore

* Forgot the comma

* Fixup makefile

* Time is an input

* Note that the actual time that applies can vary

Senders need to be at most 60s; policing and shaping needs to use at
least 60s.

* Add mention of consequences of different enforcement strategies

* Add a minimum rate rule

Rather than respecting the last received rate signal,
this makes the signals sticky for a period of time.

This makes the signals more robust.
It also makes it possible to have SCONE packets sent
by endpoints at a higher rate than network elements update them.

* twice

* Add an implementation of the advice application

* Adjust language around enforcement

It is an application (that vague amorphous thing that sits above
protocols) that receives signals, not signals, so make the language
clearer about that.

The application MAY drop state if advice is reduced.  That implies that
the network MUST allow for this to happen.  Sharpen that.

* Improve advice on sending

* Reshape advice

* This is just exemplary

* Change heading, tweaks

* Forbid SCONE packet insertion

This is one potential approach to #21.  However, it does constrain things more
than I think is necessary.  I tend to think that staying quiet is best,
because it gives network elements the option and endpoints the choice of
ignoring those inserted SCONE packets.  I wouldn't want to *encourage*
that outcome though.

Closes #21.

* Address some comments from Marcus and Alan

* Consolidate advice on following advice

* Add language about expired advice

* Applications are what enables SCONE

* Reword desync advice

Co-authored-by: Kazuho Oku <kazuhooku@gmail.com>

* these are limits

Co-authored-by: Kazuho Oku <kazuhooku@gmail.com>

* include

Co-authored-by: Kazuho Oku <kazuhooku@gmail.com>

* Better yet

* datagrams!

Co-authored-by: Marcus Ihlar <marcus@ihlar.se>

* Explain why 67

* Reword the delay piece for changing advice

* Include text about video/NAT

* oops

* less "this"

* ensures -> helps avoid

* Scope of advice

---------

Co-authored-by: Kazuho Oku <kazuhooku@gmail.com>
Co-authored-by: Marcus Ihlar <marcus@ihlar.se>