Update intro
This attempts to shorten the intro by moving some of the text that has
been added over time to the applicability section.

It also clarifies that "network element" is any function that is
on-path.  I'll add the other thing we agreed to at the last meeting to
another PR that touches the relevant text.

Closes #83.
Closes #69.