More clearly distinguish 3.1 and 3.2
The former is about the importance of congestion feedback.  The latter
is about the potential for applications to exceed the advice because
they don't know that they are sharing a scope with other network users.
For the latter, that's why networks will still need ways to monitor and
police.  The former, we already address, but it's worth making that text
clearer.